Properl initialize `sndfile` member to avoid random segmentation faults
This wold occasionally fail in the constructor. The underlying reason is that the constructor is in turn calling `openSF()` and the first thing this does is `sf_close(sndfile);`. Now, previously `sndfile` was not  initialized, so it would contain random gibberish. When the random gibberish was different from `NULL`, then the call to `sf_close(sndfile)` would cause a segmentation fault.

This patch properly initializes it to `NULL`.